acquisition/requirements-dev.txt
pytest==3.0.6
requests==2.13.0
Pillow==3.4.2
